Rename commonsense_qa to commonsense
