+++ src/benchmark/__init__.py
@@ -4,7 +4,7 @@
 from . import simple_scenarios  # noqa
 from . import mmlu_scenario  # noqa
 from . import msmarco_scenario  # noqa
-from . import commonsense_qa_scenario  # noqa
+from . import commonsense_scenario  # noqa
 from . import twitter_aae_scenario  # noqa
 from . import real_toxicity_prompts_scenario  # noqa
 from . import math_scenario  # noqa
@@ -42,7 +42,7 @@ from . import entity_data_imputation_scenario  # noqa
 
 # Metrics
 from . import basic_metrics  # noqa
-from . import commonsense_qa_metrics  # noqa
+from . import commonsense_metrics  # noqa
 from . import multiple_request_metrics  # noqa
 from . import toxicity_metrics  # noqa
 from . import tokens_metric  # noqa
