+++ src/benchmark/presentation/run_specs.conf
@@ -153,20 +153,20 @@
 
 ##### Knowledge #####
 
-"commonsense_qa:model=default_and_limited,dataset=hellaswag,method=mcqa": {status: "READY"}
-"commonsense_qa:model=default_and_limited,dataset=openbookqa,method=mcqa": {status: "READY"}
-"commonsense_qa:model=default_and_limited,dataset=commonsenseqa,method=mcqa": {status: "READY"}
-"commonsense_qa:model=default_and_limited,dataset=piqa,method=mcqa": {status: "READY"}
-"commonsense_qa:model=default_and_limited,dataset=siqa,method=mcqa": {status: "READY"}
+"commonsense:model=default_and_limited,dataset=hellaswag,method=mcqa": {status: "READY"}
+"commonsense:model=default_and_limited,dataset=openbookqa,method=mcqa": {status: "READY"}
+"commonsense:model=default_and_limited,dataset=commonsenseqa,method=mcqa": {status: "READY"}
+"commonsense:model=default_and_limited,dataset=piqa,method=mcqa": {status: "READY"}
+"commonsense:model=default_and_limited,dataset=siqa,method=mcqa": {status: "READY"}
 
 # TODO: @Michi @Percy @Dimitris @Rishi - Decide if we are doing clm strategy in main benchmark.
 # Note: We can always include clm as an ablation.
-# "commonsense_qa:dataset=hellaswag,method=clm": {status: "READY"}
-# "commonsense_qa:dataset=openbookqa,method=clm": {status: "READY"}
-# "commonsense_qa:dataset=commonsenseqa,method=clm": {status: "READY"}
+# "commonsense:dataset=hellaswag,method=clm": {status: "READY"}
+# "commonsense:dataset=openbookqa,method=clm": {status: "READY"}
+# "commonsense:dataset=commonsenseqa,method=clm": {status: "READY"}
 # TODO: Currently produces an error - see issue 223
-# "commonsense_qa:dataset=piqa,method=clm": {status: "READY"}
-# "commonsense_qa:dataset=siqa,method=clm": {status: "READY"}
+# "commonsense:dataset=piqa,method=clm": {status: "READY"}
+# "commonsense:dataset=siqa,method=clm": {status: "READY"}
 
 # For MMLU, we sampled the following 10 subjects, which cover diverse topics across humanities, social sciences and STEM.
 "mmlu:model=default_and_limited,subject=abstract_algebra": {status: "READY"}
