+++ src/benchmark/run_specs.py
@@ -14,7 +14,7 @@ from .run_expander import RUN_EXPANDERS
 from .runner import RunSpec
 from .scenario import ScenarioSpec
 
-from .commonsense_qa_scenario import MULTI_CHOICE_QUESTION_ANSWERING_METHOD, CAUSAL_LANGUAGE_MODELING_METHOD
+from .commonsense_scenario import MULTI_CHOICE_QUESTION_ANSWERING_METHOD, CAUSAL_LANGUAGE_MODELING_METHOD
 from .msmarco_scenario import MSMARCOScenario
 from .numeracy_scenario import get_numeracy_adapter_spec, RELTYPE_INFO
 from .raft_scenario import get_raft_instructions
@@ -63,8 +63,8 @@ def get_bbq_metrics() -> List[MetricSpec]:
     ]
 
 
-def get_commonsense_qa_metrics(args: Dict[str, Any]) -> List[MetricSpec]:
-    return [MetricSpec(class_name="benchmark.commonsense_qa_metrics.CommonSenseQAMetric", args=args)]
+def get_commonsense_metrics(args: Dict[str, Any]) -> List[MetricSpec]:
+    return [MetricSpec(class_name="benchmark.commonsense_metrics.CommonSenseMetric", args=args)]
 
 
 def get_msmarco_metrics(task: str, track: str, qrels_path: str, topk: Optional[int] = None) -> List[MetricSpec]:
@@ -356,10 +356,9 @@ def get_wikifact_spec(k: str, subject: str) -> RunSpec:
     )
 
 
-def get_commonsense_qa_spec(dataset: str, method: str) -> RunSpec:
+def get_commonsense_spec(dataset: str, method: str) -> RunSpec:
     scenario = ScenarioSpec(
-        class_name="benchmark.commonsense_qa_scenario.CommonSenseQAScenario",
-        args={"dataset": dataset, "method": method,},
+        class_name="benchmark.commonsense_scenario.CommonSenseScenario", args={"dataset": dataset, "method": method,},
     )
 
     if method == MULTI_CHOICE_QUESTION_ANSWERING_METHOD:
@@ -377,7 +376,7 @@ def get_commonsense_qa_spec(dataset: str, method: str) -> RunSpec:
             stop_sequences=["\n"],
         )
         run_spec = RunSpec(
-            name=f"commonsense_qa:dataset={dataset},method={method}",
+            name=f"commonsense:dataset={dataset},method={method}",
             scenario=scenario,
             adapter_spec=adapter_spec,
             metrics=get_basic_metrics({"names": ["exact_match"]}),
@@ -398,13 +397,13 @@ def get_commonsense_qa_spec(dataset: str, method: str) -> RunSpec:
             temperature=0.0,
         )
         run_spec = RunSpec(
-            name=f"commonsense_qa:dataset={dataset},method={method}",
+            name=f"commonsense:dataset={dataset},method={method}",
             scenario=scenario,
             adapter_spec=adapter_spec,
-            metrics=get_commonsense_qa_metrics({"n_choice": n_choice}),
+            metrics=get_commonsense_metrics({"n_choice": n_choice}),
         )
     else:
-        raise ValueError(f"Unknown commonsense QA method: {method}")
+        raise ValueError(f"Unknown commonsense method: {method}")
 
     return run_spec
 
@@ -1347,7 +1346,7 @@ CANONICAL_RUN_SPEC_FUNCS: Dict[str, Callable[..., RunSpec]] = {
     "mmlu": get_mmlu_spec,
     "msmarco": get_msmarco_spec,
     "narrative_qa": get_narrativeqa_spec,
-    "commonsense_qa": get_commonsense_qa_spec,
+    "commonsense": get_commonsense_spec,
     "lsat_qa": get_lsat_qa_spec,
     "quac": get_quac_spec,
     "wikifact": get_wikifact_spec,
